# src/main/resources/swagger.yaml
swagger: '2.0'
info:
  description: Api Documentation
  version: '1.0'
  title: Api Documentation
  termsOfService: 'urn:tos'
  contact: {}
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0'
host: 'localhost:8080'
basePath: /
tags:
  - name: interaction-controller
    description: Interaction Controller
paths:
  /api/interactions:
    get:
      tags:
        - interaction-controller
      summary: getInteractions
      operationId: getInteractionsUsingGET
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Interaction'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
    post:
      tags:
        - interaction-controller
      summary: getInteractions
      operationId: getInteractionsUsingPOST
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: inx
          description: inx
          required: true
          schema:
            $ref: '#/definitions/Interaction'
      responses:
        '200':
          description: OK
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  '/api/interactions/{id}':
    get:
      tags:
        - interaction-controller
      summary: getInteraction
      operationId: getInteractionUsingGET
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Interaction'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
    put:
      tags:
        - interaction-controller
      summary: getInteractions
      operationId: getInteractionsUsingPUT
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int32
        - in: body
          name: inx
          description: inx
          required: true
          schema:
            $ref: '#/definitions/Interaction'
      responses:
        '200':
          description: OK
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
definitions:
  Interaction:
    type: object
    properties:
      id:
        type: integer
        format: int32
      queue:
        type: string
      tenantId:
        type: integer
        format: int32
    title: Interaction